// Frontend/src/app/modules/ct-client/model/product-response.ts
/**
 * ProductManagementAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryDto } from './category-dto';


export interface ProductResponse { 
    id?: string;
    name?: string | null;
    description?: string | null;
    sku?: string | null;
    price?: number;
    quantity?: number;
    categoryId?: string | null;
    createdAt?: Date;
    updatedAt?: Date;
    categoryHierarchy?: Array<CategoryDto> | null;
    readonly inStock?: boolean;
}